+++ corehq/apps/es/tests/__init__.py
@@ -1,0 +1,7 @@
+from .test_aggregations import *
+from .test_esquery import *
+from .test_esqueryset import *
+from .test_facets import *
+from .test_filters import *
+from .test_queries import *
+
