+++ corehq/apps/es/tests/test_esquery.py
@@ -1,0 +1,118 @@
+from unittest import TestCase
+
+from corehq.apps.es import filters
+from corehq.apps.es import forms, users
+from corehq.apps.es.es_query import HQESQuery
+from corehq.apps.es.tests import ElasticTestMixin
+from corehq.elastic import SIZE_LIMIT
+
+
+class TestESQuery(ElasticTestMixin, TestCase):
+    maxDiff = 1000
+
+    def test_basic_query(self):
+        json_output = {
+            "query": {
+                "filtered": {
+                    "filter": {
+                        "and": [
+                            {"match_all": {}}
+                        ]
+                    },
+                    "query": {"match_all": {}}
+                }
+            },
+            "size": SIZE_LIMIT
+        }
+        self.checkQuery(HQESQuery('forms'), json_output)
+
+    def test_query_size(self):
+        json_output = {
+            "query": {
+                "filtered": {
+                    "filter": {
+                        "and": [
+                            {"match_all": {}}
+                        ]
+                    },
+                    "query": {"match_all": {}}
+                }
+            },
+            "size": 0
+        }
+        # use `is not None`; 0 or 1000000 == 1000000
+        self.checkQuery(HQESQuery('forms').size(0), json_output)
+        json_output['size'] = 123
+        self.checkQuery(HQESQuery('forms').size(123), json_output)
+
+    def test_form_query(self):
+        json_output = {
+            "query": {
+                "filtered": {
+                    "filter": {
+                        "and": [
+                            {"term": {"doc_type": "xforminstance"}},
+                            {"not": {"missing":
+                                {"field": "xmlns"}}},
+                            {"not": {"missing":
+                                {"field": "form.meta.userID"}}},
+                        ]
+                    },
+                    "query": {"match_all": {}}
+                }
+            },
+            "size": SIZE_LIMIT
+        }
+        query = forms.FormES()
+        self.checkQuery(query, json_output)
+
+    def test_user_query(self):
+        json_output = {
+            "query": {
+                "filtered": {
+                    "filter": {
+                        "and": [
+                            {"term": {"is_active": True}},
+                            {"term": {"base_doc": "couchuser"}},
+                        ]
+                    },
+                    "query": {"match_all": {}}
+                }
+            },
+            "size": SIZE_LIMIT
+        }
+        query = users.UserES()
+        self.checkQuery(query, json_output)
+
+    def test_filtered_forms(self):
+        json_output = {
+            "query": {
+                "filtered": {
+                    "filter": {
+                        "and": [
+                            {"term": {"domain.exact": "zombocom"}},
+                            {"term": {"xmlns.exact": "banana"}},
+                            {"term": {"doc_type": "xforminstance"}},
+                            {"not": {"missing":
+                                {"field": "xmlns"}}},
+                            {"not": {"missing":
+                                {"field": "form.meta.userID"}}},
+                        ]
+                    },
+                    "query": {"match_all": {}}
+                }
+            },
+            "size": SIZE_LIMIT
+        }
+        query = forms.FormES()\
+                .filter(filters.domain("zombocom"))\
+                .xmlns('banana')
+        self.checkQuery(query, json_output)
+
+    def test_remove_all_defaults(self):
+        # Elasticsearch fails if you pass it an empty list of filters
+        query = (users.UserES()
+                 .remove_default_filter('not_deleted')
+                 .remove_default_filter('active'))
+        filters = query.raw_query['query']['filtered']['filter']['and']
+        self.assertTrue(len(filters) > 0)
